% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{Calculate Centrality Measure}
\usage{
centrality(numbers)
}
\arguments{
\item{numbers}{A matrix or data frame of numeric values.}

\item{maxs}{A vector defining upper bounds for each column of `numbers`.}

\item{mins}{A vector defining lower bounds for each column of `numbers`.}
}
\value{
A matrix of centrality measures corresponding to each element in `numbers`.
}
\description{
This function calculates a centrality measure based on input numbers, maxs, and mins. TThe central3 measure quantifies the relative position of each element in numbers by assessing its deviation from the midpoint of its range defined by maxs and mins, normalized across the column's range.
}
