% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_digit.R
\name{count_digit}
\alias{count_digit}
\title{Count Occurrences of a Target Digit in a Data Frame}
\usage{
count_digit(input, target_digit)
}
\arguments{
\item{input}{A data frame. The data frame whose elements will be searched for the target digit.}

\item{target_digit}{A numeric or character value. The digit to count within the elements of the data frame.}
}
\value{
A data frame of the same dimensions as the input, where each element represents the count of the target digit in the corresponding element of the input data frame.
}
\description{
This function counts the occurrences of a specified digit in each element of a data frame.
}
\examples{
df <- data.frame(a = c(123, 456, 789), b = c(321, 654, 987))
count_digit(df, 2)
count_digit(df, '5')

}
