% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNGbiases.R
\name{RNGbiases}
\alias{RNGbiases}
\title{Calculates the presence of simplex number biases in Random Number Generation tasks}
\usage{
RNGbiases(data, mins, maxs, nsim = 10000, seed = 123456789)
}
\arguments{
\item{data}{A numeric matrix or data frame containing the individual random number selections (rows = individuals, columns = tasks)}

\item{mins}{A numeric vector specifying the lower limit of each random number generation task (e.g. 1, if the task is 1-100)}

\item{maxs}{A numeric vector specifying the lower limit of each random number generation task (e.g. 100, if the task is 1-100)}

\item{nsim}{Number of simulations to perform. Default is 10,000}

\item{seed}{Seed for reproducibility of simulations. Default is 123456789.}
}
\value{
A list containing:
  - `compositeZ`: A composite score indicating a person's propensity to simplex number bias (standardized scores compared to simulated data)
  - `rawbias`: Raw bias percentages for seven typical biases
  - `biasZ`: Standardized bias measures (Z scores for seven typical biases. Mean and SD are derived from simulated dataset).
  - `primes`: Matrix indicating prime numbers in the original dataset.
  - `even`: Matrix indicating even numbers in the original dataset.
  - `mult5`: Matrix indicating multiples of 5 in the original dataset.
  - `mult10`: Matrix indicating multiples of 10 in the original dataset.
  - `c0`: Matrix indicating the presence of digit 0 in the original dataset.
  - `c5`: Matrix indicating the presence of digit 5 in the original dataset.
  - `Central`: Matrix indicating the centrality of each value in the original dataset.
}
\description{
This function simulates datasets as a benchmark for true randomness.
}
\details{
This function generates simulated datasets (`sim1M` and `sim1Mwithoutmissing`) where random number
selection is uniform but introduces biases similar to those observed in the original dataset.
Biases are quantified using functions such as `is_prime2`, `is_multiple`, `count_digit`, and `centrality`.
The function calculates raw bias percentages (`rawbiasperc`), standardizes biases (`biaspercZ`),
and computes composite bias scores (`compositebiasZ`).
}
\examples{
data <- RNGopenpsychometrics
numbers <- data[,2:21]
mins <- c(1,1,44,100,1,123,1,7,1,60,80,1,5,600,1,600,1,1,77,30)
maxs <- c(100,1000,99,500,100,1234,100,50,100,140,90,100,8,800,100,800,4,100,4012,60)
bias <- RNGbiases(numbers, mins, maxs)

}
\seealso{
`is_prime2`, `is_multiple`, `count_digit`, `centrality`
}
